requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==24.3.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
branca==0.8.1
cachetools==5.5.0
captcha==0.6.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
entrypoints==0.4
extra-streamlit-components==0.1.71
Faker==35.0.0
favicon==0.7.0
folium==0.19.4
fonttools==4.55.7
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
htbuilder==0.9.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
narwhals==1.22.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
prometheus_client==0.21.1
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.2
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
shapely==2.0.7
six==1.17.0
smmap==5.0.2
soupsieve==2.6
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.41.1
streamlit-authenticator==0.4.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_folium==0.24.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
xyzservices==2024.9.0
